+++ src/components/chat/chat-row.tsx
@@ -10,8 +10,9 @@ import { useChatContext } from "@/features/chat/chat-ui/chat-context"
 import { ChatRole, ChatSentiment, FeedbackType, PromptMessage } from "@/features/chat/models"
 import { showError } from "@/features/globals/global-message-store"
 import { AI_NAME } from "@/features/theme/theme-config"
-import AssistantButtons from "@/features/ui/assistant-buttons"
+import { AssistantButtons, FleschButton } from "@/features/ui/assistant-buttons"
 import Modal from "@/features/ui/modal"
+import { calculateFleschScore } from "@/features/chat/chat-services/chat-flesch"
 
 interface ChatRowProps {
   chatMessageId: string
@@ -63,6 +64,8 @@ export const ChatRow: FC<ChatRowProps> = props => {
         break
     }
   }
+
+  const handleFleschIconClick = () => {}
 
   const handleCopyButton = (): void => {
     toggleButton("CopyButton")
@@ -120,6 +123,8 @@ export const ChatRow: FC<ChatRowProps> = props => {
     closeModal?.()
   }
 
+  const fleshScore = calculateFleschScore(props.message.content) ?? 0
+
   const safetyWarning = props.message.contentFilterResult ? (
     <div
       className="mt-2 flex max-w-none justify-center space-x-2 rounded-md bg-alert p-2 text-sm text-primary md:text-base"
@@ -148,11 +153,14 @@ export const ChatRow: FC<ChatRowProps> = props => {
           <Typography variant="h3" className="mt-0 flex-1 text-heading" tabIndex={0}>
             {props.name}
           </Typography>
-          {process.env.NODE_ENV === "development" && (
-            <Typography variant="h3" className="mt-0 flex-1 text-heading" tabIndex={0}>
-              {props.chatMessageId}
-            </Typography>
-          )}
+          <Typography variant="h3" className="mt-0 flex-1 text-center text-heading" tabIndex={0}>
+            {props.chatMessageId}
+          </Typography>
+          <div className="flex items-center">
+            {props.showAssistantButtons && (
+              <FleschButton handleFleschIconClick={handleFleschIconClick} fleschScore={fleshScore} />
+            )}
+          </div>
           <Modal
             chatThreadId={props.chatThreadId}
             chatMessageId={props.chatMessageId}
@@ -175,6 +183,7 @@ export const ChatRow: FC<ChatRowProps> = props => {
         <div className="sr-only" aria-live="assertive">
           {feedbackMessage}
         </div>
+
         {props.type === "assistant" && props.showAssistantButtons && (
           <AssistantButtons
             isIconChecked={isIconChecked}